requirements.txt
numpy==1.16.4
tensorflow==1.14.0
Pillow==6.2.1
